Add files via upload
Corrected many many more of my typos made while transcribing it from the Creative Computing magazine.  The file chipmunk.bas is now playable in Chipmunk Basic.  source.bas is the original which would only work on CDC Cyber BASIC.  In chipmunk.bas, all code for the CDC has been preserved, but commented out with a REM statement at the beginning of the line.